+++ migrations/000032_create_order_shippings_table.up.sql
@@ -2,9 +2,9 @@ CREATE TABLE IF NOT EXISTS order_shippings (
   id bigserial PRIMARY KEY,
   order_id bigint REFERENCES orders ON DELETE CASCADE,
   logistic_id bigint REFERENCES logistics ON DELETE CASCADE,
-  subtotal integer NOT NULL,
+  subtotal bigint NOT NULL,
   voucher_id bigint REFERENCES vouchers ON DELETE CASCADE,
-  total integer NOT NULL,
+  total bigint NOT NULL,
   created_at timestamp(0) with time zone NOT NULL DEFAULT NOW(),
   updated_at timestamp(0) with time zone NOT NULL DEFAULT NOW()
 );
